Norm and fixes

+++ ft_atof.c
@@ -6,30 +6,30 @@
 /*   By: tlernoul <tlernoul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/18 13:56:11 by tlernoul          #+#    #+#             */
-/*   Updated: 2020/06/23 12:18:37 by tlernoul         ###   ########.fr       */
+/*   Updated: 2020/07/02 09:57:09 by tlernoul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double		decimal(const char *str, int i, double ret)
+float		decimal(const char *str, int i, float ret)
 {
-	unsigned long 	dec;
+	unsigned long	dec;
 
 	dec = 1;
 	i++;
 	while (ft_isdigit(str[i]))
 	{
 		dec *= 10;
-		ret += (double)(str[i++] - '0') / (double)dec;
+		ret += (float)(str[i++] - '0') / (float)dec;
 	}
 	return (ret);
 }
 
-double		ft_atof(const char *str)
+float		ft_atof(const char *str)
 {
-	double			ret;
-	double			sign;
+	float			ret;
+	float			sign;
 	int				i;
 
 	i = 0;
@@ -39,14 +39,14 @@ double		ft_atof(const char *str)
 		return (0.0f);
 	while (!ft_isdigit(str[i]))
 	{
-        if (str[i] == '-')
-            sign = -1.0f;
-        i++;
-    }
+		if (str[i] == '-')
+			sign = -1.0f;
+		i++;
+	}
 	while (ft_isdigit(str[i]))
 	{
 		ret *= 10;
-		ret += (double)(str[i++] - '0');
+		ret += (float)(str[i++] - '0');
 	}
 	if (str[i] == '.')
 		ret = decimal(str, i, ret);
